Added: deleteUsers, Roles

+++ src/users/users.repository.ts
@@ -3,6 +3,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Users } from './entities/user.entity';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
+import { SupabaseService } from 'src/supabase/supabase.service';
+import { Role } from 'src/auth/enum/roles.enum';
 // import { UpdateUserDto } from './dto/update-user.dto';
 
 @Injectable()
@@ -10,6 +12,7 @@ export class UsersRepository {
   constructor(
     @InjectRepository(Users)
     private readonly usersRepository: Repository<Users>,
+    private supabaseService: SupabaseService,
   ) {}
 
   async getUsers(): Promise<Users[]> {
@@ -17,7 +20,7 @@ export class UsersRepository {
   }
 
   async getUserById(id: string): Promise<Users> {
-    const user = await this.usersRepository.findOneBy({ id });
+    const user = await this.usersRepository.findOne({ where: { id } });
     if (!user) {
       throw new NotFoundException(`User with ID ${id} not found`);
     }
@@ -35,10 +38,42 @@ export class UsersRepository {
   //   return this.findOne(id);
   // }
 
+  async updateRole(id: string, role: Role): Promise<Users> {
+    const user = await this.usersRepository.findOne({ where: { id } });
+
+    if (!user) {
+      throw new NotFoundException(`Usuario con ID ${id} no encontrado`);
+    }
+
+    user.role = role;
+    return this.usersRepository.save(user);
+  }
+
   async deleteUser(id: string): Promise<void> {
-    const result = await this.usersRepository.delete(id);
-    if (result.affected === 0) {
-      throw new NotFoundException(`User with ID ${id} not found`);
+    try {
+      const user = await this.usersRepository.findOne({ where: { id } });
+      if (!user) {
+        throw new NotFoundException(`Usuario con ID ${id} no encontrado`);
+      }
+      const { error } = await this.supabaseService
+        .getClient()
+        .auth.admin.deleteUser(id);
+      if (error && !error.message.includes('User not found')) {
+        throw new Error(
+          `Error al eliminar usuario de Supabase Auth: ${error.message}`,
+        );
+      }
+      const result = await this.usersRepository.delete(id);
+      if (result.affected === 0) {
+        throw new NotFoundException(`Error al eliminar usuario de PostgreSQL`);
+      }
+    } catch (error) {
+      if (error instanceof NotFoundException) {
+        throw error;
+      }
+      throw new Error(
+        `Error durante la eliminación del usuario: ${error.message}`,
+      );
     }
   }
 }
